Move plot worker into its own module, no changes except visibility modifiers
